python/powerpoint_rag_service/pyproject.toml
[tool.poetry]
name = "powerpoint-rag-service"
version = "0.1.0"
description = "A service for PowerPoint RAG using Vertex AI Vector Search"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
google-cloud-aiplatform = { extras = [
    "prediction", # For embedding models
    "matching-engine", # For vector search
], version = "^1.93.0" }
google-cloud-storage = "^2.10.0"
python-pptx = "^0.6.21"
python-dotenv = "^1.0.1" # For reading .env file in the service

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"